chore(deps): upgrade k8s.io/gengo to v2 (#9)
Migrate k8s.io/gengo -> k8s.io/gengo/v2

This required copying the extractCommentTags function because it has been
deprecated in v2. v2 introduces a new ExtractFunctionStyleCommentTags
function, but it has a different interface, so it's just simpler to
import the original ExtractCommentTags code which is dependency free.

The code generated with this tool in
github.com/oauth2-proxy/oauth2-proxy/pkg/apis/options is not affected.